DOCS Clearly differentiate readMembers and paginatedReadMembers queries
* Add GraphQL syntax highlighting to readme.
